==> climate/9. explore_timeseries_prec.Rmd <==
---
title: "Explore precipitation timeseries"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r, include = FALSE}
library(dplyr)
library(lubridate)
library(reshape)
library(ggplot2)

di <- "/Volumes/Data/Climatic_Data/EnviDat/envicloud/chelsa/chelsa_V1/climregions_tables/"
```

### PRECIPITATION ### 
```{r, echo = FALSE}
# Load time series
files <- list.files(di, pattern = "*_pr_*")
all_data <- data.frame()
for (i in 1:length(files)) {
  name <- files[i]
  model <- strsplit(name, split = "_")[[1]][1]
  period <- strsplit(name, split = "_")[[1]][3]
  temp <- read.csv(paste(di, files[i], sep = ""), sep = " ") %>%
    mutate(Model = model,
           Period = period,
           date = as.Date(paste(Year, Month, "15", sep = "-"), tryFormats = c("%Y-%m-%d")),
           month_start_date = ymd(paste(Year, Month, "01", sep = "/")) ,
           next_month_start_date = month_start_date + months(1),
           nr_days_in_month = as.integer(as.duration(interval(month_start_date, next_month_start_date)) / 86400),
           Prec_reg1_monthsum = prec_region_1*86400*nr_days_in_month,
           Prec_reg2_monthsum = prec_region_2*86400*nr_days_in_month,
           Prec_reg3_monthsum = prec_region_3*86400*nr_days_in_month,
           Prec_reg4_monthsum = prec_region_4*86400*nr_days_in_month) %>%
    dplyr::select(date, Year, Month, Model, Period, Prec_reg1_monthsum, Prec_reg2_monthsum, Prec_reg3_monthsum, Prec_reg4_monthsum)

  all_data <- rbind(all_data, temp)
}

# Fetch data
tidy_data <- all_data %>%
  dplyr::select(date, Model, Period, Prec_reg1_monthsum, Prec_reg2_monthsum, Prec_reg3_monthsum, Prec_reg4_monthsum) %>%
  melt(id.vars=c("date", "Period", "Model"))

# Subsets
ACCESS_rcp45 <- tidy_data %>%
  filter(Model == "ACCESS1-3",
         Period == "historical" | Period == "rcp45")

ACCESS_rcp85 <- tidy_data %>%
  filter(Model == "ACCESS1-3",
         Period == "historical" | Period == "rcp85")

CESM1_rcp45 <- tidy_data %>%
  filter(Model == "CESM1-BGC",
         Period == "historical" | Period == "rcp45")

CESM1_rcp85 <- tidy_data %>%
  filter(Model == "CESM1-BGC",
         Period == "historical" | Period == "rcp85")

CMCC_rcp45 <- tidy_data %>%
  filter(Model == "CMCC-CM",
         Period == "historical" | Period == "rcp45")

CMCC_rcp85 <- tidy_data %>%
  filter(Model == "CMCC-CM",
         Period == "historical" | Period == "rcp85")

MIROC5_rcp45 <- tidy_data %>%
  filter(Model == "MIROC5",
         Period == "historical" | Period == "rcp45")

MIROC5_rcp85 <- tidy_data %>%
  filter(Model == "MIROC5",
         Period == "historical" | Period == "rcp85")

# Plots
# tidy_data %>%
#   filter(variable == "Prec_reg1_monthsum",
#          date > "2090-01-15") %>%
#   ggplot(aes(x = date, y = value)) +
#   geom_point(size = 0.1) +
#   geom_line() +
#   facet_grid(Model ~ Period , scales="free_y") +
#   labs(x = "Year", y = "Monthly precipitation (mm)") +
#   ggtitle("Region 1") +
#   theme_minimal() +
#   ylim(0, 400)
# 
# tidy_data %>%
#   filter(variable == "Prec_reg2_monthsum",
#          date > "2090-01-15") %>%
#   ggplot(aes(x = date, y = value)) +
#   geom_point(size = 0.1) +
#   geom_line() +
#   facet_grid(Model ~ Period , scales="free_y") +
#   labs(x = "Year", y = "Monthly precipitation (mm)") +
#   ggtitle("Region 2") +
#   theme_minimal() +
#   ylim(0, 400)
# 
# tidy_data %>%
#   filter(variable == "Prec_reg3_monthsum",
#          date > "2090-01-15") %>%
#   ggplot(aes(x = date, y = value)) +
#   geom_point(size = 0.1) +
#   geom_line() +
#   facet_grid(Model ~ Period , scales="free_y") +
#   labs(x = "Year", y = "Monthly precipitation (mm)") +
#   ggtitle("Region 3") +
#   theme_minimal() +
#   ylim(0, 400)
# 
# tidy_data %>%
#   filter(variable == "Prec_reg4_monthsum",
#          date > "2090-01-15") %>%
#   ggplot(aes(x = date, y = value)) +
#   geom_point(size = 0.1) +
#   geom_line() +
#   facet_grid(Model ~ Period , scales="free_y") +
#   labs(x = "Year", y = "Monthly precipitation (mm)") +
#   ggtitle("Region 4") +
#   theme_minimal() +
#   ylim(0, 400)
```

```{r, echo = FALSE}
# Check year intravariability - Full period
month_mean <- all_data %>%
  select(Month, Model, Period, Prec_reg1_monthsum, Prec_reg2_monthsum, Prec_reg3_monthsum, Prec_reg4_monthsum) %>%
  group_by(Month, Model, Period) %>%
  summarise(Prec_reg1_monthsum_mean = mean(Prec_reg1_monthsum),
            Prec_reg2_monthsum_mean = mean(Prec_reg2_monthsum),
            Prec_reg3_monthsum_mean = mean(Prec_reg3_monthsum),
            Prec_reg4_monthsum_mean = mean(Prec_reg4_monthsum),
            Prec_reg1_monthsum_sd = sd(Prec_reg1_monthsum),
            Prec_reg2_monthsum_sd = sd(Prec_reg2_monthsum),
            Prec_reg3_monthsum_sd = sd(Prec_reg3_monthsum),
            Prec_reg4_monthsum_sd = sd(Prec_reg4_monthsum))

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg1_monthsum_mean - Prec_reg1_monthsum_sd, 
                  ymax = Prec_reg1_monthsum_mean + Prec_reg1_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg1_monthsum_mean)) +
  geom_point(aes(y = Prec_reg1_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 1 - Full period") +
  ylab("mean +- sd")

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg2_monthsum_mean - Prec_reg2_monthsum_sd, 
                  ymax = Prec_reg2_monthsum_mean + Prec_reg2_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg2_monthsum_mean)) +
  geom_point(aes(y = Prec_reg2_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 2 - Full period") +
  ylab("mean +- sd")

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg3_monthsum_mean - Prec_reg3_monthsum_sd, 
                  ymax = Prec_reg3_monthsum_mean + Prec_reg3_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg3_monthsum_mean)) +
  geom_point(aes(y = Prec_reg3_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 3 - Full period") +
  ylab("mean +- sd")

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg4_monthsum_mean - Prec_reg4_monthsum_sd, 
                  ymax = Prec_reg4_monthsum_mean + Prec_reg4_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg4_monthsum_mean)) +
  geom_point(aes(y = Prec_reg4_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 4 - Full period") +
  ylab("mean +- sd")

# Check year intravariability - Last decades
month_mean <- all_data %>%
  filter(Year >= 2080) %>%
  select(Month, Model, Period, Prec_reg1_monthsum, Prec_reg2_monthsum, Prec_reg3_monthsum, Prec_reg4_monthsum) %>%
  group_by(Month, Model, Period) %>%
  summarise(Prec_reg1_monthsum_mean = mean(Prec_reg1_monthsum),
            Prec_reg2_monthsum_mean = mean(Prec_reg2_monthsum),
            Prec_reg3_monthsum_mean = mean(Prec_reg3_monthsum),
            Prec_reg4_monthsum_mean = mean(Prec_reg4_monthsum),
            Prec_reg1_monthsum_sd = sd(Prec_reg1_monthsum),
            Prec_reg2_monthsum_sd = sd(Prec_reg2_monthsum),
            Prec_reg3_monthsum_sd = sd(Prec_reg3_monthsum),
            Prec_reg4_monthsum_sd = sd(Prec_reg4_monthsum))

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg1_monthsum_mean - Prec_reg1_monthsum_sd, 
                  ymax = Prec_reg1_monthsum_mean + Prec_reg1_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg1_monthsum_mean)) +
  geom_point(aes(y = Prec_reg1_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 1 - 2080-2100 period") +
  ylab("mean +- sd")

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg2_monthsum_mean - Prec_reg2_monthsum_sd, 
                  ymax = Prec_reg2_monthsum_mean + Prec_reg2_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg2_monthsum_mean)) +
  geom_point(aes(y = Prec_reg2_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 2 - 2080-2100 period") +
  ylab("mean +- sd")

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg3_monthsum_mean - Prec_reg3_monthsum_sd, 
                  ymax = Prec_reg3_monthsum_mean + Prec_reg3_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg3_monthsum_mean)) +
  geom_point(aes(y = Prec_reg3_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 3 - 2080-2100 period") +
  ylab("mean +- sd")

ggplot(data = month_mean, aes(x = Month)) +
  geom_ribbon(aes(ymin = Prec_reg4_monthsum_mean - Prec_reg4_monthsum_sd, 
                  ymax = Prec_reg4_monthsum_mean + Prec_reg4_monthsum_sd), fill = "grey70") +
  geom_line(aes(y = Prec_reg4_monthsum_mean)) +
  geom_point(aes(y = Prec_reg4_monthsum_mean), size = 0.1) +
  facet_grid(Model ~ Period) +
  theme_minimal() +
  scale_x_continuous(breaks = c(1:12)) +
  ylim(0, 200) +
  ggtitle("Region 4 - 2080-2100 period") +
  ylab("mean +- sd")
```

## Annual precipitation

```{r}
annual <- all_data %>%
  select(Year, Model, Period, Prec_reg1_monthsum, Prec_reg2_monthsum, Prec_reg3_monthsum, Prec_reg4_monthsum) %>%
  group_by(Year, Model, Period) %>%
  summarise(Prec_reg1_annual = sum(Prec_reg1_monthsum),
            Prec_reg2_annual = sum(Prec_reg2_monthsum),
            Prec_reg3_annual = sum(Prec_reg3_monthsum),
            Prec_reg4_annual = sum(Prec_reg4_monthsum))

annual_tidy <- melt(as.data.frame(annual), id.vars = c("Year", "Model", "Period"))

annual_tidy %>%
  filter(variable == "Prec_reg1_annual") %>%
  ggplot(aes(x = Year, colour = Period)) +
  geom_line(aes(y = value)) +
  geom_point(aes(y = value), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  # ylim(0, 200) +
  ggtitle("Region 1") +
  ylab("Annual precipitation (mm)")

annual_tidy %>%
  filter(variable == "Prec_reg2_annual") %>%
  ggplot(aes(x = Year, colour = Period)) +
  geom_line(aes(y = value)) +
  geom_point(aes(y = value), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  # ylim(0, 200) +
  ggtitle("Region 2") +
  ylab("Annual precipitation (mm)")

annual_tidy %>%
  filter(variable == "Prec_reg3_annual") %>%
  ggplot(aes(x = Year, colour = Period)) +
  geom_line(aes(y = value)) +
  geom_point(aes(y = value), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  # ylim(0, 200) +
  ggtitle("Region 3") +
  ylab("Annual precipitation (mm)")

annual_tidy %>%
  filter(variable == "Prec_reg4_annual") %>%
  ggplot(aes(x = Year, colour = Period)) +
  geom_line(aes(y = value)) +
  geom_point(aes(y = value), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  # ylim(0, 200) +
  ggtitle("Region 4") +
  ylab("Annual precipitation (mm)")
```
