==> climate/9. explore_timeseries_tasmean.Rmd <==
---
title: "Explore mean temperature timeseries"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r, echo = FALSE, include = FALSE}
library(dplyr)
library(lubridate)
library(reshape)
library(ggplot2)

di <- "/Volumes/Data/Climatic_Data/EnviDat/envicloud/chelsa/chelsa_V1/climregions_tables/"
```

### MEAN TEMPERATURE: average(Tmax, Tmin) ### 
```{r, echo = FALSE}
# Load time series
tasmin_files <- list.files(di, pattern = "*_tasmin_*")
tasmin_data <- data.frame()
for (i in 1:length(tasmin_files)) {
  name <- tasmin_files[i]
  model <- strsplit(name, split = "_")[[1]][1]
  period <- strsplit(name, split = "_")[[1]][3]
  temp <- read.csv(paste(di, tasmin_files[i], sep = ""), sep = " ") %>%
    mutate(Model = model,
           Period = period,
           date = as.Date(paste(Year, Month, "15", sep = "-"), tryFormats = c("%Y-%m-%d")),
           Tmin_reg1_C = tmin_region_1 - 273.15,
           Tmin_reg2_C = tmin_region_2 - 273.15,
           Tmin_reg3_C = tmin_region_3 - 273.15,
           Tmin_reg4_C = tmin_region_4 - 273.15) %>%
    dplyr::select(date, Year, Month, Model, Period, Tmin_reg1_C, Tmin_reg2_C, Tmin_reg3_C, Tmin_reg4_C)

  tasmin_data <- rbind(tasmin_data, temp)
}

tasmax_files <- list.files(di, pattern = "*_tasmax_*")
tasmax_data <- data.frame()
for (i in 1:length(tasmax_files)) {
  name <- tasmax_files[i]
  model <- strsplit(name, split = "_")[[1]][1]
  period <- strsplit(name, split = "_")[[1]][3]
  temp <- read.csv(paste(di, tasmax_files[i], sep = ""), sep = " ") %>%
    mutate(Model = model,
           Period = period,
           date = as.Date(paste(Year, Month, "15", sep = "-"), tryFormats = c("%Y-%m-%d")),
           Tmax_reg1_C = tmax_region_1 - 273.15,
           Tmax_reg2_C = tmax_region_2 - 273.15,
           Tmax_reg3_C = tmax_region_3 - 273.15,
           Tmax_reg4_C = tmax_region_4 - 273.15) %>%
    dplyr::select(date, Year, Month, Model, Period, Tmax_reg1_C, Tmax_reg2_C, Tmax_reg3_C, Tmax_reg4_C)

  tasmax_data <- rbind(tasmax_data, temp)
}
```

```{r, echo = FALSE}
reg1 <- merge(tasmin_data %>% select(date, Year, Month, Model, Period, Tmin_reg1_C), tasmax_data %>% select(date, Year, Month, Model, Period, Tmax_reg1_C)) %>%
  mutate(Tavg_C = (Tmin_reg1_C + Tmax_reg1_C)/2) %>%
  select(Year, Model, Period, Tavg_C) %>%
  group_by(Year, Model, Period) %>%
  summarise(Tmean_C_annual = mean(Tavg_C))

ggplot(reg1, aes(x = Year, colour = Period)) +
  geom_line(aes(y = Tmean_C_annual)) +
  geom_point(aes(y = Tmean_C_annual), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  ylim(0, 20) +
  ggtitle("Region 1") +
  ylab("Mean annual temperature (ºC)")
```

```{r, echo = FALSE}
reg2 <- merge(tasmin_data %>% select(date, Year, Month, Model, Period, Tmin_reg2_C), tasmax_data %>% select(date, Year, Month, Model, Period, Tmax_reg2_C)) %>%
  mutate(Tavg_C = (Tmin_reg2_C + Tmax_reg2_C)/2) %>%
  select(Year, Model, Period, Tavg_C) %>%
  group_by(Year, Model, Period) %>%
  summarise(Tmean_C_annual = mean(Tavg_C))

ggplot(reg2, aes(x = Year, colour = Period)) +
  geom_line(aes(y = Tmean_C_annual)) +
  geom_point(aes(y = Tmean_C_annual), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  ylim(0, 20) +
  ggtitle("Region 2") +
  ylab("Mean annual temperature (ºC)")
```

```{r, echo = FALSE}
reg3 <- merge(tasmin_data %>% select(date, Year, Month, Model, Period, Tmin_reg3_C), tasmax_data %>% select(date, Year, Month, Model, Period, Tmax_reg3_C)) %>%
  mutate(Tavg_C = (Tmin_reg3_C + Tmax_reg3_C)/2) %>%
  select(Year, Model, Period, Tavg_C) %>%
  group_by(Year, Model, Period) %>%
  summarise(Tmean_C_annual = mean(Tavg_C))

ggplot(reg3, aes(x = Year, colour = Period)) +
  geom_line(aes(y = Tmean_C_annual)) +
  geom_point(aes(y = Tmean_C_annual), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  ylim(0, 20) +
  ggtitle("Region 3") +
  ylab("Mean annual temperature (ºC)")
```

```{r, echo = FALSE}
reg4 <- merge(tasmin_data %>% select(date, Year, Month, Model, Period, Tmin_reg4_C), tasmax_data %>% select(date, Year, Month, Model, Period, Tmax_reg4_C)) %>%
  mutate(Tavg_C = (Tmin_reg4_C + Tmax_reg4_C)/2) %>%
  select(Year, Model, Period, Tavg_C) %>%
  group_by(Year, Model, Period) %>%
  summarise(Tmean_C_annual = mean(Tavg_C))

ggplot(reg4, aes(x = Year, colour = Period)) +
  geom_line(aes(y = Tmean_C_annual)) +
  geom_point(aes(y = Tmean_C_annual), size = 0.1) +
  facet_grid(Model ~ .) +
  theme_minimal() +
  ylim(0, 20) +
  ggtitle("Region 4") +
  ylab("Mean annual temperature (ºC)")
```
